+++ client/form.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset='utf-8'>
+  <meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1">
+  <title>Wildeer - Wild Code School</title>
+</head>
+
+<body>
+  <section id='scoreForm'>
+    <form id="form">
+      <label for="inputName">Player ID:
+        <input id="inputName" type="text" name="playerId">
+      </label>
+      <label for="inputScore">Score:
+        <input id="inputScore" type="text" name="score">
+      </label>
+      <input type="submit" value="Enregistrer">
+    </form>
+  </section>
+
+  <script type="module" src='./js/form.js'></script>
+</body>
+
+</html>
