+++ client/index.html
@@ -56,22 +56,6 @@
     <br/>
   </section>
 
-  <!-- Formulaire Score -->
-
-  <section id='scoreForm'>
-
-    <form id="form">
-      <label for="inputName">Player ID:
-        <input id="inputName" type="text" name="playerId">
-      </label>
-      <label for="inputScore">Score:
-        <input id="inputScore" type="text" name="score">
-      </label>
-      <input type="submit" value="Enregistrer">
-    </form>
-
-  </section>
-
   <!-- FOOTER -->
   <footer>
     <div class='info_footer'>
@@ -81,8 +65,6 @@
 </div>
 
   <script type="module" src='./js/app.js'></script>
-  <script type="text/javascript" src="./js/wildeer.js"></script>
-
 </body>
 
 </html>
