+++ server/database/users/4.json
@@ -1,6 +1,6 @@
 {
   "id": 4,
   "userName": "Cricri",
-  "bestScore": 5643,
+  "bestScore": 56,
   "score": []
 }
