+++ server/database/users/6.json
@@ -1,6 +1,6 @@
 {
   "id": 6,
   "userName": "Kinkin",
-  "bestScore": 56436778,
+  "bestScore": 564,
   "score": []
 }
