+++ server/database/users/8.json
@@ -1,0 +1,6 @@
+{
+  "id": 8,
+  "userName": "DEMO",
+  "bestScore": 42,
+  "score": []
+}
